+++ git_machete/config.py
@@ -43,6 +43,7 @@ class PRDescriptionIntroStyle(ParsableEnum):
 
 class MacheteConfig:
 
+    _ADVICE_MACHETE_CREATE_FROM_FORK = 'advice.macheteCreateFromFork'
     _ADVICE_MACHETE_EDITOR_SELECTION = 'advice.macheteEditorSelection'
     _SQUASH_MERGE_DETECTION = 'machete.squashMergeDetection'
     _STATUS_EXTRA_SPACE_BEFORE_BRANCH_NAME = 'machete.status.extraSpaceBeforeBranchName'
@@ -59,7 +60,10 @@ class MacheteConfig:
         self._git: Git = git if git is not None else Git()
 
     def advice_machete_editor_selection(self) -> bool:
-        return self._git.get_config_attr_or_none(self._ADVICE_MACHETE_EDITOR_SELECTION) != 'false'
+        return self._git.get_boolean_config_attr(key=self._ADVICE_MACHETE_EDITOR_SELECTION, default_value=True)
+
+    def advice_machete_create_from_fork(self) -> bool:
+        return self._git.get_boolean_config_attr(key=self._ADVICE_MACHETE_CREATE_FROM_FORK, default_value=True)
 
     def core_editor(self) -> Optional[str]:
         return self._git.get_config_attr_or_none("core.editor")